requirements.txt
runway
ccxt
openai
python-dotenv
fastapi
uvicorn
